textEditor, 파일첨부, 첨부파일목록(게시판상세화면) 컴포넌트화
