--- server/uploads/notice/10/notice_1_651645_1704040551201_0_과제2.txt
@@ -1,17 +1,0 @@
-바닐라 자바스크립트는 제이쿼리등 외부 라이브러리나 프레임워크를 사용하지 않는 순수 자바스크립트를 의미.
-이러한 표현을 쓰는 이유는 자바스크립트 버전이 업되면서 라이브러리나 프레임워크에 의존하지않고 순수 자바스크립트 실력이 중요해져가는것을 의미하는듯. 다만 생산성을 생각한다면 배제하긴 어렵다.
-
-
-var vs let vs const차이점
-
-var : 중복선언가능, 재할당가능
-let  : 중복선언불가능, 재할당가능
-const : 중복선언x, 재할당x
-
-1. 스코프
-var : 함수레벨 스코프. -> 함수내에서 블록안에 있는 var변수를 블록밖에서 접근가능
-let/const : 블록레벨 스코프
-2. 호이스팅
-var/let/const 모두 호이스팅(밑에 선언된 변수가 위에 정의된거)이 동작
-var는 변수선언과 동시에 undefined로 초기화.
-let/const는 변수선언은 해두고 초기화는 실제 변수 선언한 곳에서 초기화.
