--- server/uploads/notice/10/notice_1_651645_1704040551201_1_과제3.txt
@@ -1,12 +1,0 @@
-DOM (Document Object Model)
-- DOM은 웹 페이지의 HTML요소를 트리 자료구조 형태로 구성한 것을 말하며 웹 브라우저가 HTML문서를 읽으면 DOM 트리가 만들어지고, 자바스크립트 언어를 통해서 DOM의 요소(HTML 요소)에 접근해 조작이 가능.
-- 
-DOM요소 종류
-- 문서노드(document Node) : 트리의 루트노드에 해당되며 DOM에 접근하기 위한 시작점
-- 요소노드 : 특정 HTML 요소를 의미하는 노드. 해당 요소노드에는 어트리뷰트 노드, 텍스트 노드를 포함하고 있기에 요소를 통해 속성 및 텍스트의 조작도 가능
-
-이벤트 버블링
- - 특정 요소에 addEventListener('click'..)와 같이 특정 이벤트를 감지하는 이벤트 핸들러가 동작할 때 이벤트가 발생된 요소부터 부모 요소를 거슬러 올라가 이벤트를 전파하는 것을 의미.
- - 부모요소를 거슬러 올라가면서 요소에 할당된 이벤트 핸들러가 호출됨. (해당 요소에 이벤트 핸들러가 없다면 호출되지 않고 그 다음 부모 요소로 계속 전파)
- - 거의 모든 이벤트는 버블링이 발생하지만 예외인 경우도 있음(focus등). 또한 stopPropagation과 같은 함수를 통해서 버블링을 강제로 중단 시키는것도 가능.
-
